backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
google-generativeai==0.8.0
kaggle==1.6.6
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0
python-multipart==0.0.6
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LangChain for orchestration
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13

# Database options
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Firebase (optional)
firebase-admin==6.3.0

# Data Science & ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
joblib==1.3.2
nbformat==5.9.2

# Testing
pytest==7.4.3
